Added additional fields to models Added all sales page Added my items for sale page Added my purchases page Added my sales page Added totals on sales,cart and my purcchases page Added restrictions to Create/Update/Delete functions depending on situation. Added error pages when restrictions are encountered Added quantity control for sales items

+++ SalesBoard/Controllers/ItemsController.cs
@@ -43,6 +43,7 @@ namespace SalesBoard.Controllers
         // GET: Items/myItems
         public ActionResult MyItems()
         {
+          
             var seller = _userManager.GetUserName(User);
             var items = _context.Items
                 .Where(m => m.Seller == seller);
@@ -71,6 +72,13 @@ namespace SalesBoard.Controllers
         // GET: Items/Create
         public IActionResult Create()
         {
+            var user = _userManager.GetUserName(User);
+
+            if (user == null)
+            {
+                ViewBag.errorMessage = "You are not logged in, log in or register to create an item for sale";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
             return View();
         }
 
@@ -81,6 +89,14 @@ namespace SalesBoard.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Description,Price,Quantity")] Items items)
         {
+            var user = _userManager.GetUserName(User);
+
+            if (user == null)
+            {
+                ViewBag.errorMessage = "You are not logged in, log in or register to create an item for sale";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
             if (ModelState.IsValid)
             {
                 var seller = _userManager.GetUserName(User);
@@ -95,16 +111,30 @@ namespace SalesBoard.Controllers
         // GET: Items/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
 
             var items = await _context.Items.FindAsync(id);
+
+            var user = _userManager.GetUserName(User);
+            var seller = items.Seller;
+
+            if (user == null)
+            {
+                ViewBag.errorMessage = "You are not logged in, log in or register to start shopping";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
+            if (user!=seller)
+            {
+                ViewBag.errorMessage = "You are not the seller of this item, therefore you cannot edit this item";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
+
             if (items == null)
             {
                 return NotFound();
             }
+
             return View(items);
         }
 
@@ -113,11 +143,21 @@ namespace SalesBoard.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Description,Quantity,Price")] Items items)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Description,Quantity,Price,Seller")] Items items)
         {
+
+            var user = _userManager.GetUserName(User);
+            var seller = items.Seller;
+
             if (id != items.Id)
             {
                 return NotFound();
+            }
+
+            if (user != seller)
+            {
+                ViewBag.errorMessage = "You are not the seller of this item, therefore you cannot edit this item";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
             }
 
             if (ModelState.IsValid)
@@ -146,13 +186,22 @@ namespace SalesBoard.Controllers
         // GET: Items/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            var items = await _context.Items.FindAsync(id);
+
+            var user = _userManager.GetUserName(User);
+            var seller = items.Seller;
+
+            if (user != seller)
+            {
+                ViewBag.errorMessage = "You are not the seller of this item, therefore you cannot delete this item";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var items = await _context.Items
-                .FirstOrDefaultAsync(m => m.Id == id);
             if (items == null)
             {
                 return NotFound();
@@ -167,6 +216,16 @@ namespace SalesBoard.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var items = await _context.Items.FindAsync(id);
+            var user = _userManager.GetUserName(User);
+            var seller = items.Seller;
+
+            if (user != seller)
+            {
+                ViewBag.errorMessage = "You are not the seller of this item, therefore you cannot delete this item";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
+            
             _context.Items.Remove(items);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -180,13 +239,28 @@ namespace SalesBoard.Controllers
         // GET: Items/Purchase/5
         public async Task<IActionResult> Purchase(int? id)
         {
+            var items = await _context.Items.FindAsync(id);
+            var user = _userManager.GetUserName(User);
+            var seller = items.Seller;
+
+            if (user == seller)
+            {
+                ViewBag.errorMessage = "You are the seller of this item, therefore you cannot buy this item";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
+            if (user == null)
+            {
+                ViewBag.errorMessage = "You are not logged in, log in or register to start shopping";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var items = await _context.Items
-                .FirstOrDefaultAsync(m => m.Id == id);
+
             if (items == null)
             {
                 return NotFound();
@@ -198,7 +272,7 @@ namespace SalesBoard.Controllers
         // POST: Items/Purchase/5
         [HttpPost, ActionName("Purchase")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> PurchaseConfirmed([Bind("Item,Name,Seller,Quantity")] Cart cart, Sales sale)
+        public async Task<IActionResult> PurchaseConfirmed([Bind("Item,Name,Seller,Quantity,Price")] Cart cart, Sales sale)
         {
 
             // get or create a cart id
@@ -217,6 +291,18 @@ namespace SalesBoard.Controllers
                 return NotFound();
             }
 
+            if (items.Quantity < sale.Quantity)
+            {
+                ViewBag.errorMessage = "We do not have enough stock for your order.";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
+            if (items.Quantity == 0)
+            {
+                ViewBag.errorMessage = "Sorry, we are currently running out of stock.";
+                return View("Views/Home/Error.cshtml", ViewBag.errorMessage);
+            }
+
             //get the name of item sale
             var name = items.Name;
             cart.Name = name;
@@ -224,6 +310,10 @@ namespace SalesBoard.Controllers
             //get name of the seller 
             var seller = items.Seller;
             cart.Seller = seller;
+
+            var cost = items.Price * cart.Quantity;
+            cart.Price = cost;
+
 
             // make the sale/save to cart DB
             _context.Add(cart);
@@ -255,7 +345,7 @@ namespace SalesBoard.Controllers
         }
 
 
-
+
